Get rid of puts
gcc is smart enough to replace all instances of printf where
there is no format argument and the format string ends in a newline.
This is not good for us because puts means another imported symbol
from libc. We are on our way to get rid all of the imports.